+++ lib/src/extensions/scroll_extensions.dart
@@ -3,21 +3,29 @@ import 'package:flutter/material.dart';
 extension ScrollExtension on ScrollController {
   /// Scroll to top with given animation duration
   Future<void> animToTop({milliseconds = 300}) async {
-    this.animateTo(0,
-        duration: Duration(milliseconds: milliseconds),
-        curve: Curves.decelerate);
+    this.animateTo(
+      0,
+      duration: Duration(milliseconds: milliseconds),
+      curve: Curves.decelerate,
+    );
   }
 
   /// Scroll to bottom with given animation duration
   Future<void> animToBottom({milliseconds = 300}) async {
-    this.animateTo(this.position.maxScrollExtent,
-        duration: Duration(milliseconds: milliseconds), curve: Curves.linear);
+    this.animateTo(
+      this.position.maxScrollExtent,
+      duration: Duration(milliseconds: milliseconds),
+      curve: Curves.linear,
+    );
   }
 
   /// Scroll to given position with given animation duration
   Future<void> animateToPosition(double position, {milliseconds = 300}) async {
-    this.animateTo(position,
-        duration: Duration(milliseconds: milliseconds), curve: Curves.linear);
+    this.animateTo(
+      position,
+      duration: Duration(milliseconds: milliseconds),
+      curve: Curves.linear,
+    );
   }
 
   /// Scroll to top without animation