+++ lib/src/utils/constants.dart
@@ -9,6 +9,13 @@ const currencyEuro = '€';
 
 const playStoreBaseURL = 'https://play.google.com/store/apps/details?id=';
 const appStoreBaseURL = 'https://apps.apple.com/in/app/';
+const facebookBaseURL = 'https://www.facebook.com/';
+const instagramBaseURL = 'https://www.instagram.com/';
+const linkedinBaseURL = 'https://www.linkedin.com/in/';
+const twitterBaseURL = 'https://twitter.com/';
+const youtubeBaseURL = 'https://www.youtube.com/';
+const redditBaseURL = 'https://reddit.com/r/';
+const telegramBaseURL = 'https://t.me/';
 
 const SELECTED_LANGUAGE_CODE = 'selected_language_code';
 const THEME_MODE_INDEX = 'theme_mode_index';